The name resolver works again.
In the previous commit I printed the TCP port as a string, while it is
an integer. This caused the name resolver to crash. Sorry.
This commit also changes the name resolver so that it can be terminated
through a sigin, and not through the pressing of any character on the
keyboard.