Export OVS DPDK mempools for both sockets if present
This commit also containes a move to a single element of qemu command
line argument for the whole IVSHMEM, instead of doing separately the
mempool(s), then the rings (and this using separate ivshmem metadata).
It is unknown if this is actually needed or not. The big problem was the
fact that only mempool on socket 0 (ovs_mp_1500_0) was shared with the VM
when the OVS setup was actually using socket 1, and hence ovs_mp_1500_1
mempool.

Additional note: This worked with OVS commit 37d9e8c971a47a4.. (Sep 2016)
and QEMU 2.2.1 patched for IVSHMEM support.