Identification of ports is now mostly based on the port id that appears in NF-FG or NF description
That is the 2 in "my_vnf:2" from the NF-FG. These values must now match port ids used in
the NF description when specifying a port type (ivshmem, usvhost, etc.).

Previously, the mapping was done based on the order of the ports in the NF-FG and the index
from the NF description. This could have led to inconsistencies.

This commit only adapts the OVS-DPDK switch manager plugin, and the Libvirt and Docker compute
controller plugins.