rest_server: Implement attach/detach interfaces
the rest server exposes two new urls to allow to attach of detach a device
from a virtual machine.
The attach request must contain a port that is used to identify the
destination vm, the type (ivshmem, pci_assign), the id of the device
(it is used later on when detacching) and the device itself, cmdline in the
case of ivshmem and pci_address in the host on case of pci_assign.
The server answers with the pci address of the device in the guest.
The detach request must contain the id used when attaching.