Implemented diff of graph.
To update the graph (with new VNFs, end points), you must send the old graph plus the new things. In fact, the orchestrator calculate
the differnce by itself (before this commit, the orchestrator expected the diff).

Instead, to delete things, you can only delete the entire graph or one flow at a time, as it was before this commit.