utils/logger: add UN_LOG macro
Current logger function receives a lot of parameters, it makes that the calls
to this are very long and not easy to read.

This macro receives just the logging level and the log message itself.
In order to use this, it is necessary to define inside each .c or .cc file
a variable called LOG_MODULE_NAME, that will be use to print the module name.

Example: static const char LOG_MODULE_NAME[] = "KVM-Manager";