[Code cleanup] Removed old code that prevented the deletion of flows in particular situations.
In the old NF-FG the (internal) end point had a master graph, who was the graph defining such an endpoint.
Then, a flow of the master couldn't be deleted if it was the last flow on that endpoint, and some other graph still used that endpoint
(I know, it is a mess).