Started to remove information related to the VNF from the highlevel graph.
During the parsing of the NF-FG, no information associated to VNFs is stored in the highlevel graph.
The rest of the code however still use calls to the highlevel graph to retrieve configuration of VNF ports, env variables and control ports.

Moreover, it seems that this cleanup brings to the creation of a number of port per VNF that corresponds to the specified by the VNF itself in the
NF-FG. Instead, before the number of ports actually created corresponded to the number of ports actually used in the big-switch section.

More info in the next commits :)