Implemented internal endpoint + fixed number of virtual links.
Internal endpoint: each internal endpoint is represented with an LSI that, for each port, has an action "NORMAL".
Each graph using an internal endpoint triggers the creation of a virtual link connecting the LSI-0 to the LSI representing
the specific internal endpoint. Then, in the translation fof the graph, the internal endpoint is represented by a port of LSI-0,
as it was a physical port.

Limitations of the current code: the internal endpoint is not deleted, the internal endpoint is not considered in the update of a graph.

Virtual links: the number of virtual links required to implement an NF-FG is the little  as possible. Each VNF port and gre-tunnel endpoint in
an action require a virtual link from LSI-0 to tenant LSI. Each physical port and itnernal endpoint require a virtual link in the opposite
direction. Each virtual link can be used once in both directions.